+++ src/main/kotlin/ru/yandex/money/gradle/plugins/release/PreReleaseRotateVersionTask.kt
@@ -4,11 +4,13 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.logging.Logger
 import org.gradle.api.logging.Logging
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 import ru.yandex.money.gradle.plugins.release.changelog.ChangelogManager
-import ru.yandex.money.gradle.plugins.release.git.GitReleaseManager
+import ru.yandex.money.gradle.plugins.release.git.GitManager
 import ru.yandex.money.gradle.plugins.release.version.GradlePropertyVersionManager
 import ru.yandex.money.gradle.plugins.release.version.ReleaseInfoStorage
+import ru.yandex.money.tools.git.GitSettings
 import java.io.File
 
 /**
@@ -19,6 +21,9 @@ open class PreReleaseRotateVersionTask : DefaultTask() {
     companion object {
         private val log: Logger = Logging.getLogger(PreReleaseRotateVersionTask::class.java)
     }
+
+    @get:Input
+    lateinit var gitSettings: GitSettings
 
     private fun preReleaseByChangelog(
         changelog: File,
@@ -34,7 +39,7 @@ open class PreReleaseRotateVersionTask : DefaultTask() {
 
     @TaskAction
     fun rotateVersion() {
-        val uncommittedChanges = GitReleaseManager(project.rootDir).getUncommittedChanges()
+        val uncommittedChanges = GitManager(project.rootDir, gitSettings).getUncommittedChanges()
         if (!uncommittedChanges.isEmpty()) {
             throw GradleException("There are uncommitted changes \n" + uncommittedChanges.joinToString("\n"))
         }